Fix sucessors list errors

+++ client.py
@@ -4,36 +4,47 @@ from shutil import rmtree
 from utils import get_scrapper_instance
 
 
-def main(address, depth):
-    scrapper = get_scrapper_instance(address)
+def main(scrapper_address):
     path = 'urls/urls.txt'
-    fd = open(path,'r')
-    initial_urls = fd.readlines()
+    if not os.path.exists(path):
+        print(f'Error: No such file or directory: {path}')
+        return
+        
+    scrapper = get_scrapper_instance(scrapper_address)
+    if not scrapper:
+        print(f'Error: Could not connect to scrapper with address: {scrapper_address}')
+        return
+
+    fd = open(path, 'r')
+    urls = fd.readlines()
+    fd.close()
     count_url = 0
-    for url in initial_urls:
-        url = url.rstrip()
+
+    for url in urls:
+        url, depth = url.split(' ')
         try:
+            count = 0
             htmls = scrapper.get_html(url, depth)
-            if htmls:
-                count = 0
+            
             if os.path.exists(f'url{count_url}'):
                 rmtree(f'url{count_url}')
             os.makedirs(os.path.abspath(f'url{count_url}'))
+            
             for html in htmls:
-                f = open(f'url{count_url}/url{count}.html', 'w+')
-                f.write(html)
-                f.close()
+                fd = open(f'url{count_url}/url{count}.html', 'w+')
+                fd.write(html)
+                fd.close()
                 count += 1
         except:
-            pass
+            print(f'Error: There was an error getting the html of the url {url}')
+
         count_url += 1     
-    
-          
-    fd.close()
 
 
 if __name__ == '__main__':
-    if len(sys.argv) == 3:
-        main(sys.argv[1], int(sys.argv[2]))
+    if len(sys.argv) == 2:
+        main(sys.argv[1])
+    elif len(sys.argv) < 2:
+        print('Error: Missing arguments, you must enter the scrapper node address')
     else:
-        print('Error: Missing arguments')
+        print('Error: Too many arguments, you must enter only the scrapper node address')
